Moved tests from Main class to newly introduced AcceptanceTestsSuit
